Refine Order and OwnOrder
