+++ nautilus_trader/core/nautilus_pyo3.pyi
@@ -3182,6 +3182,7 @@ def update_book_with_trade_tick(book: OrderBook, trade: TradeTick) -> None: ...
 class OwnBookOrder:
     def __init__(
         self,
+        trader_id: TraderId,
         client_order_id: ClientOrderId,
         side: OrderSide,
         price: Price,
@@ -3191,10 +3192,16 @@ class OwnBookOrder:
         status: OrderStatus,
         ts_last: int,
         ts_accepted: int,
+        ts_submitted: int,
         ts_init: int,
+        venue_order_id: VenueOrderId | None = None,
     ) -> None: ...
     @property
+    def trader_id(self) -> TraderId: ...
+    @property
     def client_order_id(self) -> ClientOrderId: ...
+    @property
+    def venue_order_id(self) -> VenueOrderId | None: ...
     @property
     def side(self) -> OrderSide: ...
     @property
@@ -3209,6 +3216,10 @@ class OwnBookOrder:
     def status(self) -> OrderStatus: ...
     @property
     def ts_last(self) -> int: ...
+    @property
+    def ts_accepted(self) -> int: ...
+    @property
+    def ts_submitted(self) -> int: ...
     @property
     def ts_init(self) -> int: ...
     def exposure(self) -> float: ...
