+++ nautilus_trader/model/orders/base.pxd
@@ -130,6 +130,8 @@ cdef class Order:
     """UNIX timestamp (nanoseconds) when the order was initialized.\n\n:returns: `uint64_t`"""
     cdef readonly uint64_t ts_accepted
     """UNIX timestamp (nanoseconds) when the order was accepted (zero unless accepted).\n\n:returns: `uint64_t`"""
+    cdef readonly uint64_t ts_submitted
+    """UNIX timestamp (nanoseconds) when the order was submitted (zero unless submitted).\n\n:returns: `uint64_t`"""
     cdef readonly uint64_t ts_last
     """UNIX timestamp (nanoseconds) when the last order event occurred.\n\n:returns: `uint64_t`"""
